refactoring
moving a few items into components in order to reduce repaint
