+++ networks/10.10.10.10.24.json
@@ -1,0 +1,1 @@
+{"type":"network","cidr_notation":"10.10.10.10/24","subnet_bits":24,"subnet_mask":"255.255.255.0","wildcard_mask":"0.0.0.255","network_address":"10.10.10.0","broadcast_address":"10.10.10.255","assignable_hosts":254,"first_assignable_host":"10.10.10.1","last_assignable_host":"10.10.10.254"}
