+++ webui/src/App.css
@@ -7,6 +7,13 @@
   background-color: unset;
 }
 
+.rs-radio-group-picker {
+  border: unset;
+  border-top: 1px solid var(--rs-border-primary);
+  border-bottom: 1px solid var(--rs-border-primary)
+
+}
+
 @font-face {
   font-family: "IBM Plex Sans";
   src: url("./fonts/ibmplexsans/IBMPlexSans-Regular.ttf") format("truetype");
